requirements.txt
langchain
langchain-core
langchain-community

langchain-groq

langgraph

python-dotenv
grandalf